# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Noisia Public Reporting API
  version: "2.0.0"
  summary: Public reporting endpoints for customers and external integrations.
  description: |
    Public API for published Noisia Signal reports.

    V1 exposes flat tabular datasets and CSV exports for BI tools.
    V2 exposes the structured client-ready Signal output for external integrations.
servers:
  - url: https://studio.noisia.ai
    description: Production
security:
  - bearerAuth: []
  - noisiaApiKey: []
...
paths:
...
  /api/data-os/signal/{workspaceId}/mentions:
    get:
      tags: [Signal Workspace V1]
      operationId: getSignalWorkspaceMentions
      summary: Page canonical sanitized mentions using stable ordering and an opaque cursor.
      description: Returns the server-resolved read scope. The client cannot choose a population id. Governed mode emits one canonical root per mention; historical alias ids resolve to that root in detail/evidence navigation.
      security: [{ kindeSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/SignalView" }
        - { $ref: "#/components/parameters/SignalMetricKey" }
        - { $ref: "#/components/parameters/SignalStart" }
        - { $ref: "#/components/parameters/SignalEnd" }
        - { $ref: "#/components/parameters/SignalTimezone" }
        - { $ref: "#/components/parameters/SignalGranularity" }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, maximum: 100000, default: 0 }
        - name: sort
          in: query
          schema: { type: string, enum: [published, platform, conversation_role, engagement], default: published }
        - name: direction
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200": { $ref: "#/components/responses/SignalWorkspaceSuccess" }
        "304": { description: ETag still current. }
        "400": { $ref: "#/components/responses/SignalBackendError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SignalBackendError" }
        "503": { $ref: "#/components/responses/SignalBackendError" }
  /api/data-os/signal/{workspaceId}/lineage:
    get:
      tags: [Signal Workspace V1]
      operationId: getSignalWorkspaceLineage
      summary: Get basic metric definition and data watermark lineage.
      security: [{ kindeSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/SignalView" }
        - { $ref: "#/components/parameters/SignalStart" }
        - { $ref: "#/components/parameters/SignalEnd" }
        - { $ref: "#/components/parameters/SignalTimezone" }
        - { $ref: "#/components/parameters/SignalGranularity" }
        - name: metric_key
          in: query
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/SignalWorkspaceSuccess" }
        "304": { description: ETag still current. }
        "400": { $ref: "#/components/responses/SignalBackendError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SignalBackendError" }
        "503": { $ref: "#/components/responses/SignalBackendError" }
...
  /api/data-os/signal/{workspaceId}/topics-narratives/{kind}/{termKey}:
    get:
      tags: [Signal Workspace V1]
      operationId: getSignalTaxonomyTermDetail
      summary: Get one active topic or narrative with series, comparison and related terms.
      security: [{ kindeSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/SignalView" }
        - { $ref: "#/components/parameters/SignalTaxonomyKind" }
        - { $ref: "#/components/parameters/SignalTaxonomyTermKey" }
        - { $ref: "#/components/parameters/SignalStart" }
        - { $ref: "#/components/parameters/SignalEnd" }
        - { $ref: "#/components/parameters/SignalTimezone" }
        - { $ref: "#/components/parameters/SignalGranularity" }
        - { $ref: "#/components/parameters/SignalTextSearch" }
        - { $ref: "#/components/parameters/SignalComparisonStart" }
        - { $ref: "#/components/parameters/SignalComparisonEnd" }
      responses:
        "200":
          description: Governed term detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalTaxonomyTermDetailV1" }
        "304": { description: ETag still current. }
        "400": { $ref: "#/components/responses/SignalBackendError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SignalBackendError" }
        "503": { $ref: "#/components/responses/SignalBackendError" }
...
  /api/data-os/signal/{workspaceId}/admin-mentions:
    get:
      tags: [Signal Semantic Review]
      operationId: listOrInspectSignalAdminMentions
      summary: List canonical mentions or inspect one operator-safe detail.
      description: The list remains compact and adds operator_summary plus relational facets counted across every canonical root in the workspace, never only the Operational population. With mention, returns canonical and dedup lineage, sanitized quality, provenance, semantic assertions and Review, Operational memberships, enrichment and T&B references. raw_metadata is never returned.
      security: [{ kindeSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { name: mention, in: query, schema: { type: string, format: uuid } }
        - { name: inclusion_status, in: query, schema: { type: string, enum: [pending, included, excluded] } }
        - { name: resolution_state, in: query, schema: { type: string, enum: [unresolved, pending, approved_by_model, approved_by_human, approved_by_policy, rejected, superseded] } }
        - { name: review_status, in: query, schema: { type: string, enum: [unresolved, pending, approved, rejected] } }
        - { name: eligibility_status, in: query, schema: { type: string, enum: [unresolved, not_eligible, candidate, eligible] } }
        - { name: minimum_quality_score, in: query, schema: { type: number, minimum: 0, maximum: 100 } }
        - { name: quality_flag, in: query, schema: { type: string, maxLength: 100 } }
        - { name: latest_governance_action, in: query, schema: { type: string, enum: [include, exclude, revert, send_to_review] } }
        - { name: provenance_source_id, in: query, schema: { type: string, format: uuid } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200": { description: Compact list or focused operator-safe detail. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: "Preferred auth header: `Authorization: Bearer <NOISIA_REPORTING_API_KEY>`."
    noisiaApiKey:
      type: apiKey
      in: header
      name: x-noisia-api-key
      description: Fallback API key header.
    kindeSession:
      type: apiKey
      in: cookie
      name: noisia_studio_session
      description: Logical Studio session; Kinde authenticates and Noisia DB authorizes workspace and brand access.
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Stable Signal workspace id, never an output id.
...
    SignalView:
      name: view
      in: query
      required: false
      description: Closed client-safe governed view. Defaults to brand. The server resolves all policy, entity, population, compilation and binding identities; clients cannot supply those authorities.
      schema:
        type: string
        enum: [brand, competition, category, all-governed]
        default: brand
...
    SignalStart:
      name: start
      in: query
      required: true
      schema: { type: string, format: date }
    SignalEnd:
      name: end
      in: query
      required: true
      schema: { type: string, format: date }
    SignalTimezone:
      name: timezone
      in: query
      required: false
      schema: { type: string, default: workspace timezone }
    SignalGranularity:
      name: granularity
      in: query
      required: false
      schema: { type: string, enum: [day, week, month], default: day }
    SignalTextSearch:
      name: text_search
      in: query
      required: false
      description: Canonical case-insensitive full-text search, maximum 200 characters. Aliases search, query and q are accepted.
      schema: { type: string, maxLength: 200 }
    SignalTaxonomyKind:
      name: kind
      in: path
      required: true
      schema: { type: string, enum: [topic, narrative] }
    SignalTaxonomyTermKey:
      name: termKey
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z][a-z0-9]*(?:_[a-z0-9]+)*$'
    SignalComparisonStart:
      name: comparison_start
      in: query
      required: false
      description: Inclusive comparison start; must be paired with comparison_end.
      schema: { type: string, format: date }
    SignalComparisonEnd:
      name: comparison_end
      in: query
      required: false
      description: Inclusive comparison end; must be paired with comparison_start.
      schema: { type: string, format: date }
    SignalMetricKey:
      name: metric_key
      in: query
      required: true
      schema: { type: string }
...
  responses:
    SignalWorkspaceSuccess:
      description: signal-backend-v1 response. Concrete shape depends on the route.
      content:
        application/json:
          schema:
            type: object
            required: [contract_version]
            properties:
              contract_version: { const: signal-backend-v1 }
            additionalProperties: true
...
    SignalBackendError:
      description: Typed signal-backend-v1 contract error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SignalBackendErrorV1" }
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: API key cannot access this report.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Report, dataset or section not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SignalBackendErrorV1:
      type: object
      required: [contract_version, error, message, details]
      properties:
        contract_version: { const: signal-backend-v1 }
        error: { type: string, enum: [invalid_filter, unsupported_dimension, stale, partial, not_available] }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    SignalTaxonomyTermDetailV1:
      type: object
      required: [contract_version, workspace_id, corpus_id, filters_hash, kind, metric_key, term, series, related_terms, coverage, state, limitations, links]
      properties:
        contract_version: { const: signal-topics-narratives-v1 }
        workspace_id: { type: string, format: uuid }
        corpus_id: { type: string, format: uuid }
        filters_hash: { type: string }
        kind: { type: string, enum: [topic, narrative] }
        metric_key: { type: string, enum: [topic.volume, narrative.volume] }
        term: { type: object, additionalProperties: true }
        series: { type: array, items: { type: object, additionalProperties: true } }
        related_terms: { type: array, items: { type: object, additionalProperties: true } }
        coverage: { type: object, additionalProperties: true }
        state: { type: string, enum: [fresh, stale, pending, partial, not_available] }
        limitations: { type: array, items: { type: string } }
        links: { type: object, additionalProperties: true }
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          examples: [missing_api_key, invalid_api_key, forbidden_output, report_not_found, server_error]
        message:
          type: string
...